Update main.go
Ok. So the main running of an elevator. Few things remain; 1. Handeling of the external orders, this incorporates the cost function and the changing of the values for the ext.buttons to make sorting easier, 2. Add another For-loop the handle the afterOrders, 3. Debug to run the thing. But I think I'll get most done today/tonight